Minor bug caused major upset.
I was missing the variable declaration in my for in loop (key in attributes),
caused me to restart project and watch a tutorial on webpack to see if I could figure out problem.
Tutorial was needed regardless but the issues was minor (let key in attribute). Punishment for taking so many
days off..